# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShieldOS Privacy Backend API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  /**
   * @minLength 3
   * @maxLength 32
   */
  username: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface RefreshBody {
  refreshToken: string;
}

export interface AuthResponse {
  userId: number;
  username: string;
  accessToken: string;
  refreshToken: string;
}

export interface User {
  id: number;
  username: string;
  createdAt: string;
  isActive: boolean;
}

export interface ProvisionDeviceBody {
  /**
   * @minLength 1
   * @maxLength 64
   */
  deviceName: string;
}

export interface Device {
  id: number;
  name: string;
  publicKey: string;
  createdAt: string;
  isActive: boolean;
  blockedCount: number;
}

export interface DeviceProvisionResponse {
  device: Device;
  /** WireGuard .conf file content */
  configContent: string;
}

export interface DeviceListResponse {
  devices: Device[];
  total: number;
}

export interface VpnStatus {
  activeDevices: number;
  totalDevices: number;
  serverUptime: string;
}

export interface CheckDomainBody {
  domain: string;
}

export interface DomainCheckResponse {
  domain: string;
  blocked: boolean;
  category: string;
  source: string;
}

export interface BlocklistEntry {
  id: number;
  domain: string;
  category: string;
  source: string;
  addedAt: string;
}

export interface BlocklistEntriesResponse {
  entries: BlocklistEntry[];
  total: number;
  page: number;
  limit: number;
}

export type AddDomainBodyCategory =
  (typeof AddDomainBodyCategory)[keyof typeof AddDomainBodyCategory];

export const AddDomainBodyCategory = {
  ads: "ads",
  tracking: "tracking",
  malware: "malware",
  social: "social",
} as const;

export interface AddDomainBody {
  domain: string;
  category?: AddDomainBodyCategory;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface BlocklistStats {
  total: number;
  byCategory: CategoryCount[];
}

export interface BlockedRequest {
  id: number;
  domain: string;
  category: string;
  timestamp: string;
  deviceId: number;
  deviceName: string;
}

export interface BlockedRequestsResponse {
  requests: BlockedRequest[];
  total: number;
  page: number;
  limit: number;
}

export interface ThreatReport {
  id: number;
  domain: string;
  category: string;
  description: string;
  votes: number;
  verified: boolean;
  reportedAt: string;
  /** 1 for upvote, -1 for downvote, 0 for no vote */
  userVote: number;
}

export interface ThreatFeedResponse {
  threats: ThreatReport[];
  total: number;
  page: number;
  limit: number;
}

export type ReportThreatBodyCategory =
  (typeof ReportThreatBodyCategory)[keyof typeof ReportThreatBodyCategory];

export const ReportThreatBodyCategory = {
  malware: "malware",
  phishing: "phishing",
  tracking: "tracking",
  adware: "adware",
  spyware: "spyware",
  ransomware: "ransomware",
  other: "other",
} as const;

export interface ReportThreatBody {
  domain: string;
  category: ReportThreatBodyCategory;
  /** @maxLength 500 */
  description: string;
}

export type VoteBodyVote = (typeof VoteBodyVote)[keyof typeof VoteBodyVote];

export const VoteBodyVote = {
  NUMBER_1: 1,
  NUMBER_MINUS_1: -1,
} as const;

export interface VoteBody {
  vote: VoteBodyVote;
}

export interface ThreatStats {
  totalReports: number;
  verifiedThreats: number;
  pendingReview: number;
  topCategories: CategoryCount[];
}

export interface DashboardSummary {
  trackersBlockedTotal: number;
  domainsInBlocklist: number;
  activeDevices: number;
  threatsDetected: number;
  blockedLast24h: number;
  topBlockedDomain: string;
}

export interface BlockedChartDataPoint {
  hour: string;
  count: number;
}

export interface BlockedChartResponse {
  data: BlockedChartDataPoint[];
}

export interface CategoryBreakdownItem {
  category: string;
  count: number;
  percentage: number;
}

export interface CategoryBreakdownResponse {
  data: CategoryBreakdownItem[];
}

export type ListCustomBlocklistParams = {
  category?: ListCustomBlocklistCategory;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListCustomBlocklistCategory =
  (typeof ListCustomBlocklistCategory)[keyof typeof ListCustomBlocklistCategory];

export const ListCustomBlocklistCategory = {
  ads: "ads",
  tracking: "tracking",
  malware: "malware",
  social: "social",
  all: "all",
} as const;

export type ListBlockedRequestsParams = {
  hours?: number;
  page?: number;
  limit?: number;
};

export type GetThreatFeedParams = {
  category?: string;
  verified?: boolean;
  page?: number;
  limit?: number;
};