# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShieldOS Privacy Backend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshBody"
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and invalidate refresh token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshBody"
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # VPN
...
  /v1/blocklist/check:
    post:
      operationId: checkDomain
      tags: [blocklist]
      summary: Check if a domain is blocked
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckDomainBody"
      responses:
        "200":
          description: Domain check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainCheckResponse"
  /v1/blocklist/stats:
    get:
      operationId: getBlocklistStats
      tags: [blocklist]
      summary: Get blocklist stats (totals, by category)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blocklist statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlocklistStats"
  /v1/blocklist/custom:
    get:
      operationId: listCustomBlocklist
      tags: [blocklist]
      summary: List user's custom blocked domains
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [ads, tracking, malware, social, all]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Custom blocklist entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlocklistEntriesResponse"
    post:
      operationId: addCustomDomain
      tags: [blocklist]
      summary: Add a domain to user's custom blocklist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDomainBody"
      responses:
        "201":
          description: Domain added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlocklistEntry"
  /v1/blocklist/custom/{domain}:
    delete:
      operationId: removeCustomDomain
      tags: [blocklist]
      summary: Remove a domain from custom blocklist
      security:
        - bearerAuth: []
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/blocklist/blocked-requests:
    get:
      operationId: listBlockedRequests
      tags: [blocklist]
      summary: List recent blocked requests for the user's devices
      security:
        - bearerAuth: []
      parameters:
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Blocked requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedRequestsResponse"

  # THREATS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 8
      required: [username, password]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    RefreshBody:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [userId, username, accessToken, refreshToken]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required: [id, username, createdAt, isActive]

    # VPN
...
    CheckDomainBody:
      type: object
      properties:
        domain:
          type: string
      required: [domain]
    DomainCheckResponse:
      type: object
      properties:
        domain:
          type: string
        blocked:
          type: boolean
        category:
          type: string
        source:
          type: string
      required: [domain, blocked, category, source]
    BlocklistEntry:
      type: object
      properties:
        id:
          type: integer
        domain:
          type: string
        category:
          type: string
        source:
          type: string
        addedAt:
          type: string
          format: date-time
      required: [id, domain, category, source, addedAt]
    BlocklistEntriesResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/BlocklistEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [entries, total, page, limit]
    AddDomainBody:
      type: object
      properties:
        domain:
          type: string
        category:
          type: string
          enum: [ads, tracking, malware, social]
          default: tracking
      required: [domain]
    BlocklistStats:
      type: object
      properties:
        total:
          type: integer
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
      required: [total, byCategory]
    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required: [category, count]
    BlockedRequest:
      type: object
      properties:
        id:
          type: integer
        domain:
          type: string
        category:
          type: string
        timestamp:
          type: string
          format: date-time
        deviceId:
          type: integer
        deviceName:
          type: string
      required: [id, domain, category, timestamp, deviceId, deviceName]
    BlockedRequestsResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/BlockedRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [requests, total, page, limit]

    # THREATS
...